+++ CHANGELOG.md
@@ -92,6 +92,7 @@
 * pool: handle `close` WebSocket message ([Yuki Kishimoto])
 * pool: always close WebSocket connection when handlers terminate ([Yuki Kishimoto])
 * pool: better control over the handling of the termination request ([Yuki Kishimoto])
+* pool: ensure notification subscription in subscribe auto-close logic ([Yuki Kishimoto])
 * lmdb: use `EventBorrow` instead of `DatabaseEvent` ([Yuki Kishimoto])
 * ndb: refactor note-to-event conversion ([Yuki Kishimoto])
 * relay-builder: refactor shutdown mechanism to use `Notify` over `broadcast` ([Yuki Kishimoto])